Remove unnecessary casts.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/collections/trunk@1429892 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/collections/trie/ByteArrayKeyAnalyzer.java
@@ -105,8 +105,8 @@ public class ByteArrayKeyAnalyzer extends AbstractKeyAnalyzer<byte[]> {
             return false;
         }
         
-        int index = (int)(keyBitIndex / LENGTH);
-        int bit = (int)(keyBitIndex % LENGTH);
+        int index = keyBitIndex / LENGTH;
+        int bit = keyBitIndex % LENGTH;
         return (key[index] & mask(bit)) != 0;
     }
